src/requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Explainability
grad-cam>=1.4.8

# Utilities
tqdm>=4.65.0
tensorboard>=2.13.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.23.0
